requirements.txt
google-generativeai
PyPDF2
python-dotenv
opencv-python
deepface
requests
flask
tf-keras
numpy
